+++ scripts/build.sh
@@ -1,18 +1,15 @@
 #!/bin/bash
 
 build() {
-    echo 'building react'
+    echo 'building react production build'
 
-    rm -rf dist/*
+    rm -rf build/*
 
     export INLINE_RUNTIME_CHUNK=false
     export GENERATE_SOURCEMAP=false
 
     npm run prep-publish
     npm run build
-
-    mkdir -p dist
-    cp -r build/* dist
 }
 
 build